refactor: move waitUntil into Utils folder
